Move LockfreeBoundedStack and LockfreePoolingAllocator to a separate file. Call disconnectClient directly in login protocol instead of dispatching it. Get rid of unnecessary outputmessage checks.
