+++ src/outputmessage.h
@@ -102,103 +102,15 @@ class OutputMessagePool
 
 		static OutputMessage_ptr getOutputMessage();
 
-		void addProtocolToAutosend(Protocol_ptr protocol);
+		void addProtocolToAutosend(Protocol_ptr protocol) {
+			//dispatcher thread
+			bufferedProtocols.emplace_back(protocol);
+		}
 		void removeProtocolFromAutosend(const Protocol_ptr& protocol);
 	private:
 		OutputMessagePool() = default;
 		std::vector<Protocol_ptr> bufferedProtocols;
 };
 
-template <typename T, uint16_t CAPACITY>
-class LockfreeBoundedStack {
-	private:
-		struct Node
-		{
-			static constexpr uint16_t END = std::numeric_limits<decltype(CAPACITY)>::max();
-			uint16_t generation = 0;
-			uint16_t index = END;
-			uint16_t nextNode = END;
-			uint16_t unused = 0;
-			operator bool () const noexcept {
-				return index != END;
-			}
-		};
-		static_assert(sizeof(Node) == sizeof(uint64_t), "Invalid node size.");
-		static_assert(CAPACITY != 0 && CAPACITY < Node::END, "Specified capacity out of range.");
-	public:
-		LockfreeBoundedStack() {
-			for (uint16_t i = 0; i < CAPACITY; ++i) {
-				nodes[i].index = i;
-				internal_push(freeNodes, nodes[i]);
-			}
-		}
-
-		LockfreeBoundedStack(const LockfreeBoundedStack&) = delete;
-		LockfreeBoundedStack& operator=(const LockfreeBoundedStack&) = delete;
-
-		bool pop(T& ret) noexcept {
-			if (auto node = internal_pop(head)) {
-				ret = std::move(getPayload(node));
-				internal_push(freeNodes, node);
-				return true;
-			}
-
-			return false; //The stack was empty
-		}
-
-		bool push(T value) noexcept {
-			if (auto node = internal_pop(freeNodes)) {
-				getPayload(node) = value;
-				internal_push(head, node);
-				return true;
-			}
-
-			return false; //Exceeded maximum capacity
-		}
-	private:
-		Node internal_pop(std::atomic<Node>& listHead) noexcept {
-			Node currentHead, nextNode;
-			currentHead = listHead.load(std::memory_order_acquire);
-			do {
-				if (!currentHead) {
-					break;
-				}
-				nextNode = getNextNode(currentHead);
-			} while (!listHead.compare_exchange_weak(currentHead, nextNode,
-								 std::memory_order_release,
-								 std::memory_order_relaxed));
-
-			return currentHead;
-		}
-
-		void internal_push(std::atomic<Node>& listHead, Node newNode) noexcept {
-			auto currentHead = listHead.load(std::memory_order_acquire);
-			newNode.generation++;
-			do {
-				newNode.nextNode = currentHead.index;
-				nodes[newNode.index] = newNode;
-			} while (!listHead.compare_exchange_weak(currentHead, newNode, 
-								 std::memory_order_release,
-								 std::memory_order_relaxed));
-		}
-
-		T& getPayload(Node node) noexcept {
-			return payloads[node.index];
-		}
-
-		Node getNextNode(Node node) const noexcept {
-			if (node.nextNode == Node::END) {
-				return Node();
-			}
-			return nodes[node.nextNode];
-		}
-
-		T payloads[CAPACITY];
-		Node nodes[CAPACITY];
-
-		//TODO: Add padding to prevent these two from being on the same cache line
-		std::atomic<Node> freeNodes;
-		std::atomic<Node> head;
-};
 
 #endif
