+++ src/protocolgame.h
@@ -316,12 +316,12 @@ class ProtocolGame final : public Protocol
 
 		// Helpers so we don't need to bind every time
 		template <typename Callable, typename... Args>
-		void addGameTask(Callable function, Args... args) {
+		void addGameTask(Callable function, Args&&... args) {
 			addGameTaskInternal<false>(0, std::bind(function, &g_game, std::forward<Args>(args)...));
 		}
 		
 		template <typename Callable, typename... Args>
-		void addGameTaskTimed(uint32_t delay, Callable function, Args... args) {
+		void addGameTaskTimed(uint32_t delay, Callable function, Args&&... args) {
 			addGameTaskInternal<true>(delay, std::bind(function, &g_game, std::forward<Args>(args)...));
 		}
